+++ src/hooks/useClickOutside.ts
@@ -1,0 +1,25 @@
+import React, { useEffect } from "react";
+
+const useClickOutside = (
+  ref: React.RefObject<HTMLElement>,
+  handler: () => void,
+) => {
+  useEffect(() => {
+    const listener = (event: MouseEvent | TouchEvent) => {
+      if (!ref.current || ref.current.contains(event.target as Node)) {
+        return;
+      }
+      handler();
+    };
+
+    document.addEventListener("mousedown", listener);
+    document.addEventListener("touchstart", listener);
+
+    return () => {
+      document.removeEventListener("mousedown", listener);
+      document.removeEventListener("touchstart", listener);
+    };
+  }, [ref, handler]);
+};
+
+export default useClickOutside;
